+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Changelog
  * fix pending to replay dashboard contact messages
  * add timestamp in contact message list
  * fix assert project image files
+ * simplify CKEditor basic toolbar
 
 ##### Version 1.02.001 (2023-0618)
  * add Enterprise management