+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Changelog
  * fix assert project image files
  * simplify CKEditor basic toolbar
  * add ProjectCategory management
+ * make Project categories string lower conversion in public views
 
 ##### Version 1.02.001 (2023-0618)
  * add Enterprise management