+++ CHANGELOG.md
@@ -7,9 +7,12 @@ TODO
 Changelog
 =========
 
-##### Version 1.03.001 (WIP)
+##### Version 1.04.001 (WIP)
  * fix bugs
  * keep working
+ * fix problem with Scheduler Supervisor manager
+
+##### Version 1.03.001 (2023-06-27)
  * add dashboard widgets
  * add Google Analytics API bridge
  * fix pending to replay dashboard contact messages
@@ -19,7 +22,7 @@ Changelog
  * add ProjectCategory management
  * make Project categories string lower conversion in public views
 
-##### Version 1.02.001 (2023-0618)
+##### Version 1.02.001 (2023-06-18)
  * add Enterprise management
 
 ##### Version 1.01.001 (2023-06-16)